+++ k/run.sh
@@ -1,7 +1,0 @@
-#!/usr/bin/env bash
-
-if [ "$1" = "bash" ]; then
-  exec /bin/bash
-fi
-
-exec kubectl "$@"